seems to be working now

+++ minimc_header.h
@@ -44,5 +44,5 @@ double find_u_bin_u(int i, Input input);
 double find_E_bin_E(int i, Input input);
 void print_flux(Input input, double * flux);
 void tally( double E, Input input, double * flux, double Sigma_t );
-void abs_tally( double E, Input input, double * flux, long * Ra, double Sigma_t );
+void abs_tally( double E, Input input, double * flux, double * Ra, double Sigma_t );
 #endif
